+++ htdocs/numberwords/admin/about.php
@@ -57,7 +57,7 @@ llxHeader('','',$help_url);
 $linkback='<a href="'.DOL_URL_ROOT.'/admin/modules.php?restore_lastsearch_values=1">'.$langs->trans("BackToModuleList").'</a>';
 print_fiche_titre($langs->trans("NumberWordsSetup"),$linkback,'setup');
 
-print $langs->trans("DescNumberWords").'<br>';
+print '<span class="opacitymedium">'.$langs->trans("DescNumberWords").'</span><br>';
 print '<br>';
 
 $h=0;